+++ src/main/java/com/raoulvdberge/refinedstorage/api/autocrafting/ICraftingPatternContainer.java
@@ -6,7 +6,9 @@ import net.minecraftforge.items.IItemHandler;
 import net.minecraftforge.items.IItemHandlerModifiable;
 
 import javax.annotation.Nullable;
+
 import java.util.List;
+import java.util.UUID;
 
 /**
  * Represents a network node that contains crafting patterns.
@@ -18,9 +20,14 @@ public interface ICraftingPatternContainer {
     int getSpeedUpdateCount();
 
     /**
-     * @return the inventory that this container is facing
+     * @return the inventory that this container is connected to
      */
-    IItemHandler getFacingInventory();
+    IItemHandler getConnectedInventory();
+
+    /**
+     * @return the tile that this container is connected to
+     */
+    TileEntity getConnectedTile();
 
     /**
      * @return the tile that this container is facing
@@ -53,7 +60,17 @@ public interface ICraftingPatternContainer {
     BlockPos getPosition();
 
     /**
-     * @return true if this container is blocked, false otherwise
+     * Containers may be daisy-chained together.  If this container points to
+     * another one, gets the last container in the chain.  Otherwise, this
+     * container is the last in the chain.
+     *
+     * @return the root pattern container
+     */
+    @Nullable
+    ICraftingPatternContainer getProxyPatternContainer();
+
+    /**
+     * @return true if this container or its proxy is blocked, false otherwise
      */
     boolean isBlocked();
 
@@ -61,4 +78,9 @@ public interface ICraftingPatternContainer {
      * @param blocked whether the container should be blocked
      */
     void setBlocked(boolean blocked);
+
+    /**
+     * @param blockedOn the descendant that this container is blocked on
+     */
+    void setBlockedOn(UUID blockedOn);
 }