+++ src/main/java/com/raoulvdberge/refinedstorage/apiimpl/autocrafting/task/CraftingTask.java
@@ -670,7 +670,7 @@ public class CraftingTask implements ICraftingTask {
                             32
                         );
 
-                        if (step.getPattern().getContainer().getFacingTile() == null) {
+                        if (step.getPattern().getContainer().getConnectedTile() == null) {
                             element = new CraftingMonitorElementError(element, "gui.refinedstorage:crafting_monitor.machine_none");
                         } else if (!step.hasStartedProcessing() && !step.canStartProcessing()) {
                             element = new CraftingMonitorElementError(element, "gui.refinedstorage:crafting_monitor.machine_in_use");